Move scan into new screen
